fullCRUDForCustomerAndTransaction

+++ ColossalSounds.Data/Customer.cs
@@ -42,7 +42,7 @@ namespace ColossalSounds.Data
         }
 
         [DataType(DataType.PhoneNumber)]
-        public int PhoneNumber { get; set; }
+        public double PhoneNumber { get; set; }
 
         [DataType(DataType.EmailAddress)]
         public string Email { get; set; }
